Created StaticPageWebResourceRequestHandler
