+++ MobiFlightConnector.csproj
@@ -470,7 +470,8 @@
     <Compile Include="MobiFlight\Joysticks\Winwing\CduChar.cs" />
     <Compile Include="MobiFlight\Scripts\PythonEnvironment.cs" />
     <Compile Include="MobiFlight\Scripts\UserMessageCodes.cs" />
-    <Compile Include="MobiFlight\ThreadSafeWebView2.cs">
+    <Compile Include="MobiFlight\WebView\StaticPageWebResourceRequestHandler.cs" />
+    <Compile Include="MobiFlight\WebView\ThreadSafeWebView2.cs">
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="MobiFlight\Scripts\ChildProcessMonitor.cs" />
